refactor: 내 체크리스트 UI 분리 (#19)
