+++ src/features/mypage/components/index.ts
@@ -2,12 +2,12 @@ export { MyPageContent } from "@/features/mypage/components/MyPageContent";
 export type { MyPageContentProps } from "@/features/mypage/components/MyPageContent";
 export { MyPageActionCard } from "@/features/mypage/components/MyPageActionCard";
 export type { MyPageActionCardProps } from "@/features/mypage/components/MyPageActionCard";
-export { MyChecklistContent } from "@/features/mypage/components/MyChecklistContent";
-export type { MyChecklistContentProps } from "@/features/mypage/components/MyChecklistContent";
+export { MyChecklistContent } from "@/features/mypage/components/checklist/MyChecklistContent";
+export type { MyChecklistContentProps } from "@/features/mypage/components/checklist/MyChecklistContent";
 export { MyActivityContent } from "@/features/mypage/components/MyActivityContent";
 export type { MyActivityContentProps } from "@/features/mypage/components/MyActivityContent";
-export { MyChecklistChipSection } from "@/features/mypage/components/MyChecklistChipSection";
-export type { MyChecklistChipSectionProps } from "@/features/mypage/components/MyChecklistChipSection";
+export { MyChecklistChipSection } from "@/features/mypage/components/checklist/MyChecklistChipSection";
+export type { MyChecklistChipSectionProps } from "@/features/mypage/components/checklist/MyChecklistChipSection";
 export { MyProfileEditContent } from "@/features/mypage/components/MyProfileEditContent";
 export type { MyProfileEditContentProps } from "@/features/mypage/components/MyProfileEditContent";
 export { MyPageMenuSection } from "@/features/mypage/components/MyPageMenuSection";
